+++ NEWS.md
@@ -1,3 +1,9 @@
+# rwhatsapp 0.2.3
+
+* Updated emoji data to unicode Emoji List v13.0.
+* lookup_emoji is now an exported function.
+* Bug fixes for several time formats
+
 # rwhatsapp 0.2.2
 
 * Support for even more time formats.